requirements.txt
flask
numpy
pandas
matplotlib
langchain
openai
fastapi
google-generativeai
python-dotenv
requests
